backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15.0,<4.0
django-cors-headers>=4.4.0,<5.0
python-dotenv>=1.0.0,<2.0
requests>=2.31.0,<3.0
reportlab>=4.0.0,<5.0
Pillow>=10.0.0,<11.0
